+++ lib/plaid/client.rb
@@ -59,6 +59,11 @@ module Plaid
 
     ##
     # :attr_reader:
+    # Public: The Plaid::DepositSwitch product accessor.
+    subproduct :deposit_switch
+
+    ##
+    # :attr_reader:
     # Public: The Plaid::Identity product accessor.
     subproduct :identity
 
@@ -79,6 +84,11 @@ module Plaid
 
     ##
     # :attr_reader:
+    # Public: The Plaid::Liabilities product accessor.
+    subproduct :liabilities
+
+    ##
+    # :attr_reader:
     # Public: The Plaid::Processor product accessor.
     subproduct :processor
 
@@ -91,6 +101,21 @@ module Plaid
     # :attr_reader:
     # Public: The Plaid::Transactions product accessor.
     subproduct :transactions
+
+    ##
+    # :attr_reader:
+    # Public: The Plaid::Investments product accessor.
+    subproduct :investments
+
+    ##
+    # :attr_reader:
+    # Public: The Plaid::PaymentInitiation product accessor.
+    subproduct :payment_initiation
+
+    ##
+    # :attr_reader:
+    # Public: The Plaid::Webhooks endpoint accessor.
+    subproduct :webhooks
 
     # Public: Make a post request
     #